u8500-common: Increase max speaker volume
This fixes the low volume from speaker on incoming calls.
When the Phone app receives a call it starts ringing and during that it sets the volume for speaker from its max stream volume defined here:
http://goo.gl/odm304

That seems to happen here only:

http://goo.gl/wO82Kc

But since looks that other third party apps may take the same value from Audio Service, on AOSP code there is a prop implemented here, that will override the default value set to 5, and use the value of that prop

In our case, defined to 9 seems to work well as some users reported.